change bright into intellega

+++ forum.go
@@ -106,7 +106,7 @@ func BrowseUnitforumPost(ctx *gin.Context) {
 			unitforums[i].Choose = 0
 		} else {
 			db.Where("author = ? AND uid = ?", uauthor, unitforums[i].Id).Get(&emojiRecord)
-			unitforums[i].Choose = uint8(emojiRecord.Emoji + 1)
+			unitforums[i].Choose = emojiRecord.Emoji + 1
 		}
 	}
 	ctx.JSON(http.StatusOK, gin.H{
@@ -116,21 +116,21 @@ func BrowseUnitforumPost(ctx *gin.Context) {
 	})
 }
 
-func DBaddMainforum(title string, text string, author uint, kind uint8) {
+func DBaddMainforum(title string, text string, author int, kind int) {
 	mainforum := &Forum{Title: title, Author: author, Views: 0, Kind: kind}
 	db.Insert(mainforum)
-	unitforum := ForumComment{Text: text, Mid: uint(mainforum.Id), Author: author}
+	unitforum := ForumComment{Text: text, Mid: int(mainforum.Id), Author: author}
 	db.Insert(unitforum)
 	return
 }
 
-func DBaddUnitforum(text string, mid uint, author uint) {
+func DBaddUnitforum(text string, mid int, author int) {
 	unitforum := ForumComment{Text: text, Mid: mid, Author: author}
 	db.Insert(unitforum)
 	return
 }
 
-func DBaddEmoji(emoji uint, uid uint, author uint) {
+func DBaddEmoji(emoji int, uid int, author int) {
 	var unitforum ForumComment
 	db.ID(uid).Get(&unitforum)
 	switch emoji {
@@ -151,7 +151,7 @@ func DBaddEmoji(emoji uint, uid uint, author uint) {
 	case 7:
 		unitforum.Eyes++
 	}
-	emojiRecord := EmojiRecord{Author: author, Uid: uid, Emoji: uint8(emoji)}
+	emojiRecord := EmojiRecord{Author: author, Uid: uid, Emoji: int(emoji)}
 	db.Insert(&emojiRecord)
 	db.ID(uid).Update(&unitforum)
 	return
@@ -175,8 +175,8 @@ func AddMainforum(ctx *gin.Context) {
 	}
 	author := session.Get("userid")
 	vauthor := author.(int64)
-	uauthor := uint(vauthor)
-	DBaddMainforum(title, text, uauthor, ukind)
+	uauthor := int(vauthor)
+	DBaddMainforum(title, text, uauthor, int(ukind))
 	return
 }
 
@@ -184,10 +184,10 @@ func AddUnitforum(ctx *gin.Context) {
 	session := sessions.Default(ctx)
 	author := session.Get("userid")
 	vauthor := author.(int64)
-	uauthor := uint(vauthor)
+	uauthor := int(vauthor)
 	mid, text := ctx.PostForm("mid"), ctx.PostForm("text")
 	vmid, _ := strconv.Atoi(mid)
-	umid := uint(vmid)
+	umid := int(vmid)
 	DBaddUnitforum(text, umid, uauthor)
 	return
 }
@@ -196,12 +196,12 @@ func AddEmoji(ctx *gin.Context) {
 	session := sessions.Default(ctx)
 	author := session.Get("userid")
 	vauthor := author.(int64)
-	uauthor := uint(vauthor)
+	uauthor := int(vauthor)
 	emoji, uid := ctx.PostForm("emoji"), ctx.PostForm("uid")
 	vuid, _ := strconv.Atoi(uid)
 	vemoji, _ := strconv.Atoi(emoji)
-	uuid := uint(vuid)
-	uemoji := uint(vemoji)
+	uuid := int(vuid)
+	uemoji := int(vemoji)
 	if uemoji > 7 {
 		ctx.AbortWithStatus(http.StatusBadRequest) //传入的表情编号>7(不存在)
 		return
@@ -216,7 +216,7 @@ func FinishForum(ctx *gin.Context) {
 	vmid, _ := strconv.Atoi(mid)
 	author := session.Get("userid")
 	vauthor := author.(int64)
-	uauthor := uint(vauthor)
+	uauthor := int(vauthor)
 	var mainforum Forum
 	db.ID(vmid).Get(&mainforum)
 	if uauthor == mainforum.Author {
